requirements.txt
# Core document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Text processing
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Vector embeddings and database
sentence-transformers
chromadb
numpy>=1.24.0

# OCR and advanced document processing - Removed vllm (too heavy for Streamlit Cloud)
# pytesseract  # Commented out - requires system dependencies
Pillow>=8.0.0
# pdf2image  # Commented out - requires system dependencies

# Web interface
streamlit>=1.28.0

# AI API integration
openai>=1.0.0

# Basic utilities
tqdm>=4.66.0